[ios] Stop using reuse queue
This commit temporarily stops using the reuse queue until we have a
solution for querying point annotations on the map that is tested
for the case of querying while the map is panned, rotated, and
zoomed.

The existing solution was not performant because:

1) There was a bug that prevented the reuse queue from being
fully utilized since the updateAnnotationViews logic would pull
enqueued views right off the queue when it found a nil annotationView.

2) The constant lookup / add / remove operations on the NSArray
implementation of the reuse queue required a lot of CPU to keep up with.

Although this commit is a regression, it is actually more performant
to not use the reuse queue and to just update the center of every
annotation. Once we have the ability to query for annotations that
are visible in the current view port it'll make sense to bring
back the reuse queue.